Remove descriptions.rs
